refactor task page
