+++ src/pages/Board/utils/types.ts
