+++ src/pages/Task/components/MainContent/Activity/index.ts
@@ -1,0 +1,1 @@
+export { default as Activity } from './Activity';
