+++ src/pages/Task/components/MainContent/AddDescription/index.ts
@@ -1,0 +1,1 @@
+export { default as AddDescription } from './AddDescription';
