+++ src/theme/themes.ts
@@ -53,6 +53,7 @@ export const defaultTheme = {
       800: 'hsla(0, 0%, 100%, 0.8)',
     },
     darkness: {
+      40: 'hsla(0, 0%, 0%, 0.04)',
       50: 'hsla(0, 0%, 0%, 0.08)',
       100: 'hsla(0, 0%, 0%, 0.1)',
       200: 'hsla(0, 0%, 0%, 0.2)',
